[ANDROSDK-2007] refactor new tracker importer files
