+++ core/src/main/java/org/hisp/dhis/android/network/tracker/JobReportDTO.kt
@@ -26,7 +26,7 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.network.trackerimporter
+package org.hisp.dhis.android.network.tracker
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
