+++ core/src/main/java/org/hisp/dhis/android/network/tracker/NewRelationshipDTO.kt
@@ -26,48 +26,33 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.network.trackerimporter
+package org.hisp.dhis.android.network.tracker
 
 import kotlinx.serialization.Serializable
 import org.hisp.dhis.android.core.arch.helpers.DateUtils
-import org.hisp.dhis.android.core.trackedentity.NewTrackerImporterTrackedEntity
-import org.hisp.dhis.android.network.common.dto.BaseDeletableDataObjectDTO
-import org.hisp.dhis.android.network.common.dto.GeometryDTO
-import org.hisp.dhis.android.network.common.dto.toDto
+import org.hisp.dhis.android.core.relationship.NewTrackerImporterRelationship
 
 @Serializable
-internal data class NewTrackerImporterTrackedEntityDTO(
-    override val deleted: Boolean?,
-    val trackedEntity: String,
+internal data class NewRelationshipDTO(
+    val relationship: String,
+    val relationshipType: String?,
+    val relationshipName: String?,
     val createdAt: String?,
     val updatedAt: String?,
-    val createdAtClient: String?,
-    val updatedAtClient: String?,
-    val organisationUnit: String?,
-    val trackedEntityType: String?,
-    val geometry: GeometryDTO?,
-    val aggregatedSyncState: String?,
-    val attributes: List<NewTrackerImporterTrackedEntityAttributeValueDTO>?,
-    val enrollments: List<NewTrackerImporterEnrollmentDTO>?,
-    val programOwners: List<NewTrackerImporterProgramOwnerDTO>?,
-    val relationships: List<NewTrackerImporterRelationshipDTO>? = null,
-) : BaseDeletableDataObjectDTO
+    val bidirectional: Boolean?,
+    val from: NewRelationshipItemDTO?,
+    val to: NewRelationshipItemDTO?,
+)
 
-internal fun NewTrackerImporterTrackedEntity.toDto(): NewTrackerImporterTrackedEntityDTO {
-    return NewTrackerImporterTrackedEntityDTO(
-        deleted = this.deleted(),
-        trackedEntity = this.uid(),
+internal fun NewTrackerImporterRelationship.toDto(): NewRelationshipDTO {
+    return NewRelationshipDTO(
+        relationship = this.uid(),
+        relationshipType = this.relationshipType(),
+        relationshipName = this.relationshipName(),
         createdAt = this.createdAt()?.let { DateUtils.DATE_FORMAT.format(it) },
         updatedAt = this.updatedAt()?.let { DateUtils.DATE_FORMAT.format(it) },
-        createdAtClient = this.createdAtClient()?.let { DateUtils.DATE_FORMAT.format(it) },
-        updatedAtClient = this.updatedAtClient()?.let { DateUtils.DATE_FORMAT.format(it) },
-        organisationUnit = this.organisationUnit(),
-        trackedEntityType = this.trackedEntityType(),
-        geometry = this.geometry()?.let { it.toDto() },
-        aggregatedSyncState = this.aggregatedSyncState()?.name,
-        attributes = this.trackedEntityAttributeValues()?.map { it.toDto() },
-        enrollments = this.enrollments()?.map { it.toDto() },
-        programOwners = this.programOwners()?.map { it.toDto() },
-//        relationships = this.relationships()?.map { it.toDto() }
+        bidirectional = this.bidirectional(),
+        from = this.from()?.toDto(),
+        to = this.to()?.toDto(),
     )
 }
