+++ core/src/main/java/org/hisp/dhis/android/network/tracker/NewTrackedEntityDataValueDTO.kt
@@ -26,37 +26,31 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.network.trackerimporter
+package org.hisp.dhis.android.network.tracker
 
-import org.hisp.dhis.android.core.trackedentity.internal.NewTrackerImporterPayload
-import org.hisp.dhis.android.core.trackedentity.internal.ObjectWithUidWebResponse
-import org.hisp.dhis.android.core.tracker.importer.internal.JobProgress
-import org.hisp.dhis.android.core.tracker.importer.internal.JobReport
-import org.hisp.dhis.android.core.tracker.importer.internal.TrackerImporterNetworkHandler
-import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
-import org.koin.core.annotation.Singleton
+import kotlinx.serialization.Serializable
+import org.hisp.dhis.android.core.arch.helpers.DateUtils
+import org.hisp.dhis.android.core.trackedentity.NewTrackerImporterTrackedEntityDataValue
 
-@Singleton
-internal class TrackerImporterNetworkHandlerImpl(
-    httpServiceClient: HttpServiceClientKotlinx,
-) : TrackerImporterNetworkHandler {
-    val service = TrackerImporterService(httpServiceClient)
-    override suspend fun postTrackerPayload(
-        payload: NewTrackerImporterPayload,
-        importStrategy: String,
-    ): ObjectWithUidWebResponse {
-        val payloadDTO: NewTrackerImporterPayloadDTO = payload.toDto()
-        val apiPayload = service.postTrackerPayload(payloadDTO, ATOMIC_MODE_OBJECT, importStrategy)
-        return apiPayload.toDomain()
-    }
+@Serializable
+internal data class NewTrackedEntityDataValueDTO(
+    val event: String?,
+    val createdAt: String?,
+    val updatedAt: String?,
+    val dataElement: String?,
+    val createdBy: UserInfoDTO?,
+    val value: String?,
+    val providedElsewhere: String?,
+)
 
-    override suspend fun getJobReport(jobId: String): JobReport {
-        val apiPayload = service.getJobReport(jobId)
-        return apiPayload.toDomain()
-    }
-
-    override suspend fun getJob(jobId: String): List<JobProgress> {
-        val apiPayload = service.getJob(jobId)
-        return apiPayload.map { it.toDomain() }
-    }
+internal fun NewTrackerImporterTrackedEntityDataValue.toDto(): NewTrackedEntityDataValueDTO {
+    return NewTrackedEntityDataValueDTO(
+        event = this.event(),
+        createdAt = this.createdAt()?.let { DateUtils.DATE_FORMAT.format(it) },
+        updatedAt = this.updatedAt()?.let { DateUtils.DATE_FORMAT.format(it) },
+        dataElement = this.dataElement(),
+        createdBy = this.createdBy()?.let { it.toDto() },
+        value = this.value(),
+        providedElsewhere = this.providedElsewhere().toString(),
+    )
 }
