+++ core/src/main/java/org/hisp/dhis/android/network/tracker/NewUserInfoDTO.kt
@@ -25,48 +25,19 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.network.trackerimporter
 
-import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
-import org.koin.core.annotation.Singleton
+package org.hisp.dhis.android.network.tracker
 
-internal const val TRACKER_URL = "tracker"
-internal const val JOBS_URL = "tracker/jobs/"
+import kotlinx.serialization.Serializable
+import org.hisp.dhis.android.core.trackedentity.NewTrackerImporterUserInfo
 
-internal const val ATOMIC_MODE = "atomicMode"
-internal const val ATOMIC_MODE_OBJECT = "OBJECT"
+@Serializable
+internal data class NewUserInfoDTO(
+    val uid: String?,
+)
 
-internal const val IMPORT_STRATEGY = "importStrategy"
-internal const val IMPORT_STRATEGY_CREATE_AND_UPDATE = "CREATE_AND_UPDATE"
-internal const val IMPORT_STRATEGY_DELETE = "DELETE"
-
-@Singleton
-internal class TrackerImporterService(private val client: HttpServiceClientKotlinx) {
-
-    suspend fun postTrackerPayload(
-        payload: NewTrackerImporterPayloadDTO,
-        atomicMode: String,
-        importStrategy: String,
-    ): ObjectWithUidWebResponseDTO {
-        return client.post {
-            url(TRACKER_URL)
-            parameters {
-                attribute(ATOMIC_MODE, atomicMode)
-                attribute(IMPORT_STRATEGY, importStrategy)
-            }
-            body(payload)
-        }
-    }
-
-    suspend fun getJobReport(jobId: String): JobReportDTO {
-        return client.get {
-            url("$JOBS_URL$jobId/report")
-        }
-    }
-
-    suspend fun getJob(jobId: String): List<JobProgressDTO> {
-        return client.get {
-            url("$JOBS_URL$jobId")
-        }
-    }
+internal fun NewTrackerImporterUserInfo.toDto(): NewUserInfoDTO {
+    return NewUserInfoDTO(
+        uid = this.uid(),
+    )
 }
